alphabet-desktop/requirements.txt
# tested with Python 3.10.11
mediapipe==0.10.14
tensorflow==2.10.0
numpy==1.24.3
opencv-python==4.9.0.80
pillow==10.3.0